requirements.txt
rich~=13.8
black~=23.3
mypy~=1.4
typer~=0.13
wheel~=0.45
tomli~=2.0
